fix(issuer2): metadata service
