+++ waltid-services/waltid-issuer-api2/src/main/kotlin/id/walt/issuer2/service/openid4vci/MetadataService.kt
@@ -1,7 +1,10 @@
 package id.walt.issuer2.service.openid4vci
 
+import id.walt.crypto.keys.Key
+import id.walt.crypto.keys.KeyManager
 import id.walt.issuer2.config.Issuer2MetadataConfig
 import id.walt.issuer2.config.Issuer2ServiceConfig
+import id.walt.issuer2.service.CredentialProfileService
 import id.walt.openid4vci.metadata.issuer.CredentialConfiguration
 import id.walt.openid4vci.metadata.issuer.CredentialIssuerMetadata
 import id.walt.openid4vci.metadata.issuer.IssuerDisplay
@@ -10,11 +13,16 @@ import id.walt.sdjwt.metadata.issuer.JWTVCIssuerMetadata
 import id.walt.sdjwt.metadata.type.SdJwtVcTypeMetadataDraft04
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
+import kotlinx.serialization.json.JsonPrimitive
+import kotlinx.serialization.json.buildJsonArray
+import kotlinx.serialization.json.buildJsonObject
+import kotlinx.serialization.json.contentOrNull
+import kotlinx.serialization.json.jsonPrimitive
 
 class MetadataService(
     serviceConfig: Issuer2ServiceConfig,
     metadataConfig: Issuer2MetadataConfig,
-    private val jwksService: JwksService,
+    private val profileService: CredentialProfileService,
 ) {
     private val json = Json {
         ignoreUnknownKeys = true
@@ -24,6 +32,7 @@ class MetadataService(
 
     private val baseUrl = serviceConfig.baseUrl.trimEnd('/') + "/openid4vci"
     private val vctAuthorityBaseUrl = serviceConfig.baseUrl.trimEnd('/')
+    private val tokenSigningKeyConfig = serviceConfig.ciTokenKey
 
     private val issuerDisplay: List<IssuerDisplay>? =
         metadataConfig.issuerDisplay
@@ -74,7 +83,46 @@ class MetadataService(
 
     fun issuerBaseUrl(): String = baseUrl
 
-    suspend fun listJwks(): JsonObject = jwksService.listJwks()
+    suspend fun listJwks(): JsonObject {
+        val tokenSigningKey = KeyManager.resolveSerializedKey(tokenSigningKeyConfig)
+        val profileIssuerKeys = profileService.listProfiles()
+            .map { profile -> KeyManager.resolveSerializedKey(profile.issuerKey) }
+
+        return buildJsonObject {
+            put("keys", buildJsonArray {
+                (listOf(tokenSigningKey) + profileIssuerKeys)
+                    .map { key -> key.getPublicJwkWithKid() }
+                    .deduplicated()
+                    .forEach { add(it) }
+            })
+        }
+    }
+
+    private suspend fun Key.getPublicJwkWithKid(): JsonObject {
+        val publicJwk = getPublicKey().exportJWKObject()
+        return JsonObject(publicJwk.toMutableMap().apply {
+            putIfAbsent("kid", JsonPrimitive(getKeyId()))
+        })
+    }
+
+    private fun List<JsonObject>.deduplicated(): List<JsonObject> {
+        val seen = mutableSetOf<String>()
+        return filter { jwk ->
+            val keys = jwk.deduplicationKeys()
+            if (keys.any { it in seen }) {
+                false
+            } else {
+                seen.addAll(keys)
+                true
+            }
+        }
+    }
+
+    private fun JsonObject.deduplicationKeys(): Set<String> =
+        setOfNotNull(
+            this["kid"]?.jsonPrimitive?.contentOrNull?.takeIf { it.isNotBlank() }?.let { "kid:$it" },
+            JsonObject(filterKeys { it != "kid" }).toString().let { "jwk:$it" },
+        )
 
     private fun CredentialConfiguration.withResolvedVct(credentialType: String): CredentialConfiguration =
         if (vct == INTERNAL_VCT_BASE_URL) copy(vct = selfHostedVct(credentialType)) else this
